% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{add}
\alias{add}
\title{add a triple (subject, predicate, object) to the rdf graph}
\usage{
add(x, subject, predicate, object, subjectType = as.character(NA),
  objectType = as.character(NA), datatype_uri = as.character(NA))
}
\arguments{
\item{x}{an rdf graph object}

\item{subject}{character string containing the subject}

\item{predicate}{character string containing the predicate}

\item{object}{character string containing the object}

\item{subjectType}{the Node type of the subject, i.e. "blank", "uri"}

\item{objectType}{the Node type of the object, i.e. "blank", "uri"}

\item{datatype_uri}{the datatype URI to associate with a object literal value}
}
\value{
the rdf graph object.
}
\description{
add a triple (subject, predicate, object) to the rdf graph
}
\details{
Since the rdf graph object simply contains external pointers
to the model object in C code, note that the input object is modified
directly.
}
\examples{
x <- rdf()
add(x, 
    subject="http://www.dajobe.org/",
    predicate="http://purl.org/dc/elements/1.1/language",
    object="en")

}
