% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_methods.R
\name{c.rdf}
\alias{c.rdf}
\title{Concatenate rdf Objects.}
\usage{
\method{c}{rdf}(...)
}
\arguments{
\item{...}{objects to be concatenated}
}
\description{
All subsequent rdf objects will be appended to the first rdf object
Note: this does not free memory from any of the individual rdf objects
Note: It is generally better to avoid the use of this function by passing
an existing rdf object to and rdf_parse or rdf_add objects. Multiple active
rdf objects can cause problems when using disk-based storage backends.
}
