% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{parse}
\alias{parse}
\title{Parse RDF files}
\usage{
parse(file, format = c("rdfxml", "nquads", "ntriples", "trig", "turtle"), ...)
}
\arguments{
\item{file}{path to the rdf file to serialize}

\item{format}{rdf serialization format of the file,
one of "rdfxml", "nquads", "ntriples", "trig", or "turtle"}

\item{...}{additional parameters (not implemented)}
}
\value{
an rdf S3 object, containing the redland world
 and model objects
}
\description{
Parse RDF files
}
\examples{
file <- system.file("extdata", "dc.rdf", package="redland")
rdf <- parse(file)

}
