% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{query.rdf}
\alias{query.rdf}
\title{SPARQL query}
\usage{
query.rdf(rdf, query, ...)
}
\arguments{
\item{rdf}{an rdf object (e.g. from \code{\link{parse}})}

\item{query}{a SPARQL query, as text string}

\item{...}{additional arguments to a redland initialize-Query}
}
\value{
a list of all query results
}
\description{
SPARQL query
}
\examples{
file <- system.file("extdata", "dc.rdf", package="redland")

sparql <-
'PREFIX dc: <http://purl.org/dc/elements/1.1/>
 SELECT ?a ?c
 WHERE { ?a dc:creator ?c . }'

rdf <- parse(file)
query.rdf(rdf, sparql)

}
