% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{rdf_query}
\alias{rdf_query}
\title{SPARQL query}
\usage{
rdf_query(x, query, ...)
}
\arguments{
\item{x}{an rdf object (e.g. from \code{\link{parse}})}

\item{query}{a SPARQL query, as text string}

\item{...}{additional arguments to a redland initialize-Query}
}
\value{
a list of all query results
}
\description{
SPARQL query
}
\examples{
doc <- system.file("extdata", "dc.rdf", package="redland")

sparql <-
'PREFIX dc: <http://purl.org/dc/elements/1.1/>
 SELECT ?a ?c
 WHERE { ?a dc:creator ?c . }'

rdf <- rdf_parse(doc)
rdf_query(rdf, sparql)

}
