% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{rdf_serialize}
\alias{rdf_serialize}
\title{Serialize RDF docs}
\usage{
rdf_serialize(x, doc, format = c("rdfxml", "nquads", "ntriples", "trig",
  "turtle", "jsonld"), namespace = NULL, prefix = NULL, ...)
}
\arguments{
\item{x}{an rdf object (e.g. from \code{\link{parse}})}

\item{doc}{path to the rdf doc to parse}

\item{format}{rdf serialization format of the doc,
one of "rdfxml", "nquads", "ntriples", "trig", "turtle"
or "jsonld"}

\item{namespace}{string giving the namespace to set}

\item{prefix}{string giving the prefix associated with the namespace}

\item{...}{additional parameters (not implemented)}
}
\value{
an integer containing the return status where non
 zero indicates an error occured during serialization
}
\description{
Serialize RDF docs
}
\examples{
infile <- system.file("extdata", "dc.rdf", package="redland")
out <- tempfile("file", fileext = ".rdf")

rdf <- rdf_parse(infile)
rdf_serialize(rdf, out)

## With a namespace
rdf_serialize(rdf,
          out,
          namespace = "http://purl.org/dc/elements/1.1/",
          prefix = "dc")

}
