% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nquads.R
\name{read_nquads}
\alias{read_nquads}
\title{read an nquads file}
\usage{
read_nquads(file, ...)
}
\arguments{
\item{file}{path to nquads file}

\item{...}{additional arguments to \code{\link[=rdf_parse]{rdf_parse()}}}
}
\value{
an rdf object.  See \code{\link[=rdf_parse]{rdf_parse()}}
}
\description{
read an nquads file
}
\examples{
tmp <- tempfile(fileext = ".nq")
library(datasets)
write_nquads(iris, tmp)
read_nquads(tmp)

}
