% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{serialize}
\alias{serialize}
\title{Serialize RDF files}
\usage{
serialize(x, file, format, namespace, prefix, ...)
}
\arguments{
\item{x}{an rdf object (e.g. from \code{\link{parse}})}

\item{file}{path to the rdf file to serialize}

\item{format}{rdf serialization format of the file,
one of "rdfxml", "nquads", "ntriples", "trig", or "turtle"}

\item{namespace}{string giving the namespace to set}

\item{prefix}{string giving the prefix associated with the namespace}

\item{...}{additional parameters (not implemented)}
}
\value{
an integer containing the return status where non
 zero indicates an error occured during serialization
}
\description{
Serialize RDF files
}
\examples{
infile <- system.file("extdata", "dc.rdf", package="redland")
out <- tempfile("file", fileext = ".rdf")

rdf <- parse_rdf(infile)
serialize(rdf, out)

## With a namespace
serialize(rdf,
          out,
          namespace = "http://purl.org/dc/elements/1.1/",
          prefix = "dc")

}
