% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nquads.R
\name{write_nquads}
\alias{write_nquads}
\title{write object out as nquads}
\usage{
write_nquads(x, file, ...)
}
\arguments{
\item{x}{an object that can be represented as nquads}

\item{file}{output filename}

\item{...}{additional parameters, see examples}
}
\description{
write object out as nquads
}
\examples{
tmp <- tempfile(fileext = ".nq")
library(datasets)

## convert data.frame to nquads
write_nquads(iris, tmp)
rdf <- read_nquads(tmp)

## or starting a native rdf object
write_nquads(rdf, tempfile(fileext = ".nq"))
}
