% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{rdf_free}
\alias{rdf_free}
\title{Free Memory Associated with RDF object}
\usage{
rdf_free(rdf)
}
\arguments{
\item{rdf}{an rdf object}
}
\description{
Free Memory Associated with RDF object
}
\details{
Free all pointers associated with an rdf object.
Frees memory associated with the storage, world, and model
objects. After this a user should remove the rdf object
from the environment as well with \code{rm}, since attempting
to reference an object after it has been removed can crash
R!
}
\examples{
rdf <- rdf()
rdf_free(rdf)
rm(rdf)
}
