% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{rdf_parse}
\alias{rdf_parse}
\title{Parse RDF Files}
\usage{
rdf_parse(doc, format = c("rdfxml", "nquads", "ntriples", "turtle", "jsonld"),
  ...)
}
\arguments{
\item{doc}{path, URL, or literal string of the rdf document to parse}

\item{format}{rdf serialization format of the doc,
one of "rdfxml", "nquads", "ntriples", "turtle"
or "jsonld"}

\item{...}{additional parameters (not implemented)}
}
\value{
an rdf object, containing the redland world
and model objects
}
\description{
Parse RDF Files
}
\examples{
doc <- system.file("extdata", "dc.rdf", package="redland")
rdf <- rdf_parse(doc)

}
