% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_query.R
\name{rdf_query}
\alias{rdf_query}
\title{Perform a SPARQL Query}
\usage{
rdf_query(rdf, query, ...)
}
\arguments{
\item{rdf}{an rdf object (e.g. from \code{\link{rdf_parse}})}

\item{query}{a SPARQL query, as text string}

\item{...}{additional arguments to a redland initialize-Query}
}
\value{
a data.frame of all query results
}
\description{
Perform a SPARQL Query
}
\examples{
doc <- system.file("extdata", "dc.rdf", package="redland")

sparql <-
'PREFIX dc: <http://purl.org/dc/elements/1.1/>
 SELECT ?a ?c
 WHERE { ?a dc:creator ?c . }'

rdf <- rdf_parse(doc)
rdf_query(rdf, sparql)

}
